refactored the find missing imports functionality
