+++ ext/vesper-history/test/edu/ucsc/refactor/NavigableVesperTest.java
@@ -5,6 +5,7 @@ import edu.ucsc.refactor.internal.HostImpl;
 import edu.ucsc.refactor.internal.Upstream;
 import edu.ucsc.refactor.util.Commit;
 import edu.ucsc.refactor.util.CommitPublisher;
+import edu.ucsc.refactor.util.Recommender;
 import org.eclipse.egit.github.core.Comment;
 import org.eclipse.egit.github.core.Gist;
 import org.eclipse.egit.github.core.GistFile;
@@ -37,7 +38,7 @@ public class NavigableVesperTest extends VesperTest {
         final Set<Issue> issues = navigableRefactorer.detectIssues(CODE);
         assertThat(!issues.isEmpty(), is(true));
 
-        final List<Change> suggestedChanges = refactorer.recommendChanges(src, issues);
+        final List<Change> suggestedChanges = Recommender.recommendChanges(refactorer, src, issues);
         assertThat(suggestedChanges.isEmpty(), is(false));
 
         final Change first = suggestedChanges.get(0);
